renaming + adding tests
