+++ k-distribution/pom.xml
@@ -50,7 +50,7 @@
     </dependency>
     <dependency>
       <groupId>org.kframework.k</groupId>
-      <artifactId>kale-backend</artifactId>
+      <artifactId>skala-backend</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
