+++ k-distribution/tutorial/1_k/2_imp/lesson_1/tests/sum.imp
@@ -2,7 +2,7 @@
 // the sum of numbers from 1 to n.
 
 int n, sum;
-n = 100;
+n = 10;
 sum = 0;
 while (!(n <= 0)) {
   sum = sum + n;
